Separate SQL functions from server lobby
I also attempted to reuse some code that checks if a peer is banned
